add wallet for user - subtract commission from wallet for unit

+++ src/main/java/com/AlTaraf/Booking/Entity/unit/Unit.java
@@ -164,7 +164,7 @@ public class Unit extends Auditable<String> {
 
     private Integer price = 0;
 
-    private Integer commission = 20;
+    private Double commission = 20.0;
 
     private Double latForMapping;
